Added implicit StringContexts for standard and legacy SQL
bqs for standard and bql for legacy, these will run the queries and will get the rows as a Seq[WrappedTableRow].

The implicits are created in "Implicits" and are extended by BQSyntax. to be able to use these you need to extend your model with it and create an implicit projectId and an implicit bigquery4s.BigQuery.

It will anyway force you to create a projectId and a bigquery4s.BigQuery at the Member level. Something like this:

class Member extends BQSyntax {
  override val bq: BigQuery = BigQuery()
  override val projectId: ProjectId = ProjectId("some_project")
}